move classes to their root domain package
