+++ src/main/java/com/example/petshop/user/UserRepository.java
@@ -1,5 +1,6 @@
-package com.example.petshop.user.data;
+package com.example.petshop.user;
 
+import com.example.petshop.user.data.User;
 import org.bson.types.ObjectId;
 import org.springframework.data.mongodb.repository.MongoRepository;
