+++ src/main/java/com/example/petshop/user/controller/AuthController.java
@@ -3,7 +3,7 @@ package com.example.petshop.user.controller;
 import com.example.petshop.user.controller.dto.LoginDTO;
 import com.example.petshop.user.controller.dto.SignupDTO;
 import com.example.petshop.user.controller.dto.TokenDTO;
-import com.example.petshop.user.service.UserService;
+import com.example.petshop.user.UserService;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.AuthenticationException;
